# backend/openapi/register-user.yaml
openapi: 3.0.3
info:
  title: Voicebot user registration
  version: 1.0.0
  description: >
    Call register_user only after the user has explicitly confirmed that all
    collected details are correct.
servers:
  - url: https://YOUR_APP.azurewebsites.net
    description: Replace with BACKEND_PUBLIC_URL or your App Service hostname
paths:
  /api/users/register:
    post:
      operationId: register_user
      security:
        - ApiKeyHeader: []
        - {}
      summary: Register a confirmed user in Azure SQL
      description: >
        Persists the user in dbo.Users. Use only after explicit confirmation
        (e.g. "yes", "that's correct", "confirm").
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationResult"
        "400":
          description: Invalid input
        "409":
          description: Email already registered
        "503":
          description: Database not configured
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    UserRegistration:
      type: object
      required:
        - firstName
        - lastName
        - email
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        birthDate:
          type: string
          format: date
          nullable: true
        phoneNumber:
          type: string
          nullable: true
        streetAddress:
          type: string
          nullable: true
        zipCode:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
    RegistrationResult:
      type: object
      properties:
        userId:
          type: integer
        email:
          type: string
        createdAt:
          type: string